remove etcd dependency from broker/client and broker/protocol
