+++ broker/list_apply_api.go
@@ -10,6 +10,7 @@ import (
 	"go.etcd.io/etcd/clientv3"
 	"go.gazette.dev/core/allocator"
 	pb "go.gazette.dev/core/broker/protocol"
+	pbx "go.gazette.dev/core/broker/protocol/ext"
 	"google.golang.org/grpc/peer"
 )
 
@@ -32,7 +33,7 @@ func (svc *Service) List(ctx context.Context, req *pb.ListRequest) (resp *pb.Lis
 
 	resp = &pb.ListResponse{
 		Status: pb.Status_OK,
-		Header: pb.NewUnroutedHeader(s),
+		Header: pbx.NewUnroutedHeader(s),
 	}
 	if err = req.Validate(); err != nil {
 		return resp, err
@@ -55,9 +56,10 @@ func (svc *Service) List(ctx context.Context, req *pb.ListRequest) (resp *pb.Lis
 		},
 	}
 	for cur, ok := it.Next(); ok; cur, ok = it.Next() {
+		var journalSpec *pb.JournalSpec = s.Items[cur.Left].Decoded.(allocator.Item).ItemValue.(*pb.JournalSpec)
 		var journal = pb.ListResponse_Journal{
-			Spec: *s.Items[cur.Left].Decoded.(allocator.Item).ItemValue.(*pb.JournalSpec)}
-
+			Spec: *journalSpec,
+		}
 		metaLabels = pb.ExtractJournalSpecMetaLabels(&journal.Spec, metaLabels)
 		allLabels = pb.UnionLabelSets(metaLabels, journal.Spec.LabelSet, allLabels)
 
@@ -65,8 +67,8 @@ func (svc *Service) List(ctx context.Context, req *pb.ListRequest) (resp *pb.Lis
 			continue
 		}
 		journal.ModRevision = s.Items[cur.Left].Raw.ModRevision
-		journal.Route.Init(s.Assignments[cur.RightBegin:cur.RightEnd])
-		journal.Route.AttachEndpoints(s.KS)
+		pbx.Init(&journal.Route, s.Assignments[cur.RightBegin:cur.RightEnd])
+		pbx.AttachEndpoints(&journal.Route, s.KS)
 
 		resp.Journals = append(resp.Journals, journal)
 	}
@@ -92,7 +94,7 @@ func (svc *Service) Apply(ctx context.Context, req *pb.ApplyRequest) (resp *pb.A
 
 	resp = &pb.ApplyResponse{
 		Status: pb.Status_OK,
-		Header: pb.NewUnroutedHeader(s),
+		Header: pbx.NewUnroutedHeader(s),
 	}
 	if err = req.Validate(); err != nil {
 		return resp, err
