+++ broker/protocol/route_extensions_test.go
@@ -2,81 +2,9 @@ package protocol
 
 import (
 	gc "github.com/go-check/check"
-	"go.etcd.io/etcd/mvcc/mvccpb"
-	"go.gazette.dev/core/allocator"
-	"go.gazette.dev/core/keyspace"
 )
 
 type RouteSuite struct{}
-
-func (s *RouteSuite) TestInitialization(c *gc.C) {
-	var rt Route
-
-	rt.Init(nil)
-	c.Check(rt, gc.DeepEquals, Route{Primary: -1, Members: nil})
-
-	var kv = keyspace.KeyValues{
-		{Decoded: allocator.Assignment{MemberZone: "zone-a", MemberSuffix: "member-1", Slot: 1}},
-		{Decoded: allocator.Assignment{MemberZone: "zone-a", MemberSuffix: "member-3", Slot: 3}},
-		{Decoded: allocator.Assignment{MemberZone: "zone-b", MemberSuffix: "member-2", Slot: 2}},
-		{Decoded: allocator.Assignment{MemberZone: "zone-c", MemberSuffix: "member-4", Slot: 0}},
-	}
-	rt.Init(kv)
-
-	c.Check(rt, gc.DeepEquals, Route{
-		Primary: 3,
-		Members: []ProcessSpec_ID{
-			{Zone: "zone-a", Suffix: "member-1"},
-			{Zone: "zone-a", Suffix: "member-3"},
-			{Zone: "zone-b", Suffix: "member-2"},
-			{Zone: "zone-c", Suffix: "member-4"},
-		},
-	})
-
-	kv = kv[:3] // This time, remove the primary Assignment.
-	rt.Init(kv)
-
-	c.Check(rt, gc.DeepEquals, Route{
-		Primary: -1,
-		Members: []ProcessSpec_ID{
-			{Zone: "zone-a", Suffix: "member-1"},
-			{Zone: "zone-a", Suffix: "member-3"},
-			{Zone: "zone-b", Suffix: "member-2"},
-		},
-	})
-}
-
-func (s *RouteSuite) TestEndpointAttachmentAndCopy(c *gc.C) {
-	var ks = keyspace.KeySpace{
-		Root: "/root",
-		KeyValues: keyspace.KeyValues{
-			{Raw: mvccpb.KeyValue{Key: []byte("/root/members/zone-a#member-1")},
-				Decoded: allocator.Member{Zone: "zone-a", Suffix: "member-1",
-					MemberValue: &BrokerSpec{ProcessSpec: ProcessSpec{Endpoint: "http://host-a:port/path"}}}},
-			{Raw: mvccpb.KeyValue{Key: []byte("/root/members/zone-b#member-2")},
-				Decoded: allocator.Member{Zone: "zone-b", Suffix: "member-2",
-					MemberValue: &BrokerSpec{ProcessSpec: ProcessSpec{Endpoint: "http://host-b:port/path"}}}},
-		},
-	}
-	var rt = Route{
-		Members: []ProcessSpec_ID{
-			{Zone: "zone-a", Suffix: "member-1"},
-			{Zone: "zone-a", Suffix: "member-3"},
-			{Zone: "zone-b", Suffix: "member-2"},
-		},
-	}
-
-	rt.AttachEndpoints(&ks)
-	c.Check(rt.Endpoints, gc.DeepEquals, []Endpoint{"http://host-a:port/path", "", "http://host-b:port/path"})
-
-	var other = rt.Copy()
-
-	// Expect |other| is deeply equal while referencing different slices.
-	c.Check(rt.Members, gc.DeepEquals, other.Members)
-	c.Check(rt.Endpoints, gc.DeepEquals, other.Endpoints)
-	c.Check(&rt.Members[0], gc.Not(gc.Equals), &other.Members[0])
-	c.Check(&rt.Endpoints[0], gc.Not(gc.Equals), &other.Endpoints[0])
-}
 
 func (s *RouteSuite) TestValidationCases(c *gc.C) {
 	var rt = Route{
