+++ go.mod
@@ -8,9 +8,6 @@ require (
 	github.com/aws/aws-sdk-go v1.19.32
 	github.com/dustin/go-humanize v0.0.0-20171111073723-bb3d318650d4
 	github.com/dustinkirkland/golang-petname v0.0.0-20190613200456-11339a705ed2
-	github.com/facebookgo/ensure v0.0.0-20160127193407-b4ab57deab51 // indirect
-	github.com/facebookgo/stack v0.0.0-20160209184415-751773369052 // indirect
-	github.com/facebookgo/subset v0.0.0-20150612182917-8dac2c3c4870 // indirect
 	github.com/go-check/check v0.0.0-20180628173108-788fd7840127
 	github.com/gogo/protobuf v1.2.1
 	github.com/golang/protobuf v1.3.2
@@ -22,8 +19,8 @@ require (
 	github.com/jgraettinger/cockroach-encoding v1.1.0
 	github.com/jgraettinger/urkel v0.1.2
 	github.com/klauspost/compress v1.5.0
-	github.com/klauspost/cpuid v1.2.1 // indirect
-	github.com/kr/pretty v0.1.0 // indirect
+	github.com/klauspost/cpuid v1.2.2 // indirect
+	github.com/kr/pretty v0.2.0 // indirect
 	github.com/lib/pq v1.1.1
 	github.com/mattn/go-sqlite3 v1.11.0
 	github.com/olekukonko/tablewriter v0.0.0-20170122224234-a0225b3f23b5
@@ -36,14 +33,11 @@ require (
 	github.com/stretchr/testify v1.3.0
 	github.com/tecbot/gorocksdb v0.0.0-20190705090504-162552197222
 	go.etcd.io/etcd v0.0.0-20190711162406-e56e8471ec18
-	golang.org/x/crypto v0.0.0-20190701094942-4def268fd1a4 // indirect
 	golang.org/x/net v0.0.0-20190724013045-ca1201d0de80
 	golang.org/x/oauth2 v0.0.0-20190604053449-0f29369cfe45
 	golang.org/x/sync v0.0.0-20190423024810-112230192c58
-	golang.org/x/sys v0.0.0-20190712062909-fae7ac547cb7 // indirect
 	google.golang.org/api v0.7.0
 	google.golang.org/grpc v1.22.0
-	gopkg.in/alecthomas/kingpin.v2 v2.2.6 // indirect
 	gopkg.in/yaml.v2 v2.2.2
 	k8s.io/api v0.0.0-20190620073856-dcce3486da33
 )
